+++ halo2_gadgets/src/ecc/chip/mul_fixed/full_width.rs
@@ -187,14 +187,17 @@ pub mod tests {
     use pasta_curves::pallas;
     use rand::rngs::OsRng;
 
-    use crate::ecc::{
-        chip::{EccChip, FixedPoint as _, H},
-        tests::{FullWidth, TestFixedBases},
-        FixedPoint, NonIdentityPoint, Point, ScalarFixed,
+    use crate::{
+        ecc::{
+            chip::{EccChip, FixedPoint as _, H},
+            tests::{FullWidth, TestFixedBases},
+            FixedPoint, NonIdentityPoint, Point, ScalarFixed,
+        },
+        utilities::lookup_range_check::PallasLookupRangeCheck,
     };
 
-    pub(crate) fn test_mul_fixed(
-        chip: EccChip<TestFixedBases>,
+    pub(crate) fn test_mul_fixed<Lookup: PallasLookupRangeCheck>(
+        chip: EccChip<TestFixedBases, Lookup>,
         mut layouter: impl Layouter<pallas::Base>,
     ) -> Result<(), Error> {
         let test_base = FullWidth::from_pallas_generator();
@@ -209,18 +212,18 @@ pub mod tests {
     }
 
     #[allow(clippy::op_ref)]
-    fn test_single_base(
-        chip: EccChip<TestFixedBases>,
+    fn test_single_base<Lookup: PallasLookupRangeCheck>(
+        chip: EccChip<TestFixedBases, Lookup>,
         mut layouter: impl Layouter<pallas::Base>,
-        base: FixedPoint<pallas::Affine, EccChip<TestFixedBases>>,
+        base: FixedPoint<pallas::Affine, EccChip<TestFixedBases, Lookup>>,
         base_val: pallas::Affine,
     ) -> Result<(), Error> {
-        fn constrain_equal_non_id(
-            chip: EccChip<TestFixedBases>,
+        fn constrain_equal_non_id<Lookup: PallasLookupRangeCheck>(
+            chip: EccChip<TestFixedBases, Lookup>,
             mut layouter: impl Layouter<pallas::Base>,
             base_val: pallas::Affine,
             scalar_val: pallas::Scalar,
-            result: Point<pallas::Affine, EccChip<TestFixedBases>>,
+            result: Point<pallas::Affine, EccChip<TestFixedBases, Lookup>>,
         ) -> Result<(), Error> {
             let expected = NonIdentityPoint::new(
                 chip,